+++ rules/astro.js
@@ -1,0 +1,7 @@
+module.exports = {
+  customSyntax: "postcss-html",
+  rules: {
+    // Allow javascript custom properties (camelCase)
+    "custom-property-pattern": null,
+  },
+};